+++ PropertyInspector/CHANGELOG.md
@@ -5,11 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
- - 
 ### Added
  - 
 ### Changed
  - 
+
+## [0.1.0]
+### Added
+- Support for Space, HideInInspector, and Tooltip attributes
+### Fixed
+ - Modifying values affected by the property in the inspector by using something other than the property will no longer get overwritten by the property.
 
 ## [0.0.1] - 2022-10-24
 ### Added